pyproject.toml
[project]
name = "allaboutdjango"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django==5.2.*",
    "django-ipware>=7.0.1",
    "django-stubs==5.2.*",
    "requests>=2.32.4",
]
